[🔨Refactor] MyCharacterView 재사용성 고려하여 말풍선 입히는 MainCharacterWithBubbleView 생성 및 분리
